tetris simple for the most part is equivariant
