+++ model.py
@@ -18,8 +18,8 @@ class Model(torch.nn.Module):
 
         # first layer
         self.layer1 = Layer(self.starting_irreps_id, "5x0e + 5x1o")
-        self.activation_layer1 = ActivationLayer("GELU", "5x0e + 8x1o")
-        self.layer2 = Layer("5x0e + 8x1o", "10x0e")
+        self.activation_layer1 = ActivationLayer("GELU", "5x0e + 5x1o")
+        self.layer2 = Layer("5x0e + 5x1o", "10x0e")
         self.activation_layer2 = ActivationLayer("GELU", "10x0e")
 
         # output layer
